coding-test/receipt_ai_platform/requirements.txt
gradio==5.44.1
pillow==11.3.0
pytesseract==0.3.13
python-dotenv==1.1.1
google-genai==1.32.0